Add files via upload
These six documents help pave and design mindset that helps the ease of implementation from the ground floor up on analytics. Beginning with the Analytics Gameplan, I focus on a vision and roadmap an organization should take to implement an analytics program that is maximizes its potential and outputs quality projects for a business. Following that we have three documents that are focused on the planning and implementation of individual projects. 

The Analytics Project Plan, Analytic Planning Document and Prioritization Matrix are all documents that can help not only detail what projects are most important but can help build a project lifecycle with clear deliverables that each stakeholder in the project will be tasked with delivering. 

Finally, there are two documents that I believe are necessary for each project at the end of its build. One is the Data Quality Review, which will allow another member of the analytics team to review the work done by their colleague so they can make sure nothing was missed during the build of their project. Last, the Project Post-Mortem, where we can assess after deployment what went right, what went wrong, what was new to other implementations that we should keep, and what went wrong and needs to change before the start of the next project.